Refactor code for user testing
