+++ packages/user/src/routes/users.routes.js
@@ -7,13 +7,10 @@ const usersRouter = express.Router();
 
 //usersRouter.use(ensureAuthenticated)
 
-usersRouter.post("/", UserController.create);
-usersRouter.post("/id", UserController.getByUsername);
 usersRouter.post("/user/", UserController.sendEmail);
 usersRouter.post("/user/code/", UserController.checkCode);
 usersRouter.post("/user/password/", UserController.changePassword);
 usersRouter.post("/user/getUser/", UserController.getByToken);
 usersRouter.post("/user/edit/", UserController.editUser);
-
 
 module.exports = usersRouter;
